+++ NEWS.md
@@ -1,5 +1,26 @@
 # dials (development version)
 
+* `prop_terms()` is a new parameter object used for recipes that do supervised feature selection (#395). 
+
+# dials 1.4.1
+
+* Two new parameters, `cal_method_class()` and `cal_method_reg(),` to control which method is used to calibrate model predictions (#383).
+
+* `regularization_factor()` is now exclusive of the lower border 0 in compliance with `ranger::ranger()` (#381).
+
+
+# dials 1.4.0
+
+* For space-filling designs for $p$ parameters, there is a higher likelihood of finding a space-filling design for `1 < size <= p`. Also, single-point designs now default to a random grid (#363).
+
+* `value_seq()` and `value_sample()` now respect the `inclusive` argument of quantitative parameters (#347).
+
+* The constructors, `new_*_parameter()`, now label unlabeled parameter (i.e., constructed with `label = NULL`) as such (#349).
+
+* All messages, warnings and errors has been translated to use {cli} package (#311). 
+
+* `parameters.list()` now enforces the unused ellipsis to be empty (#378).
+
 * Added three new parameters for use in postprocessing in the tailor package (#357).
      - `buffer()` sets the distance on either side of a classification threshold 
        within which predictions are considered equivocal in 
@@ -7,7 +28,14 @@
      - `lower_limit()` and `upper_limit()` sets the ranges for
        numeric predictions in `tailor::adjust_numeric_range()`.
 
-* All messages, warnings and errors has been translated to use {cli} package (#311). 
+## Breaking changes
+
+* The `grid_*()` functions now error instead of warn when provided with the wrong argument to control the grid size. So `grid_space_filling()`, `grid_random()`, `grid_max_entropy()`, and `grid_latin_hypercube()` now error if used with a `levels` argument and `grid_regular()` now errors if used with a `size` argument (#368).
+
+* The `"optimal"` option for the `weight_func()` parameter has been removed since it is choosing the optimal value based on the resubstition error (#370). 
+
+* When constructing integer-valued parameters with a range of two consecutive values the `inclusive` argument needs to be set to `c(TRUE, TRUE)` to leave at least two values to sample from (#373). 
+
 
 # dials 1.3.0